need for work github pages
MEN'S RECIPE FOR CHARLOTTE (simple apple pie).

Take 10 eggs from the refrigerator, put the remaining seven on the table and wipe the floor. Take a large bowl and break the eggs on its edge. Pour their contents into a bowl. Wipe the table. So, you have 5 eggs in a bowl. Now take a mixer, insert the whisk and start whipping. Try inserting the corollas again, now until it clicks. Beat it up. Wash your face, neck, arms, back, and ears. As a result, you have two beaten eggs left in the bowl. Cover the walls and ceiling of the kitchen with a newspaper, cover the furniture with some cloth - it's time to get the flour ... Pour 200 g into a glass, then pour into a bowl. Carefully put the remaining 800 g back into the bag. Beat it up. Take a shower. Take 4 large apples and a knife. Prepare iodine, a patch and a bandage. Peel the apples. Treat your thumb with iodine and bandage it. Cut the apples into cubes and remember, you will need 2 apples, so you can only eat half of them during cooking. Treat your index and middle fingers with iodine. Drop the only remaining and already sliced apple into a bowl, pick up the fallen pieces from the floor, wash them and put them in a bowl too. Beat everything with a mixer. Clean the refrigerator. Now pour the contents into the pan and place in the oven. Wait an hour and, seeing no noticeable changes, turn on the oven. When you wake up, do not call the Rescue Service, but simply turn off the oven and open the windows...